+++ src/error/index.js
@@ -1,5 +1,4 @@
 import UndefinedUsed from './UndefinedUsed';
 
-export default {
-  UndefinedUsed,
-};
+export default { UndefinedUsed };
+